This patch to HTML.pm adds the ability for people to shut off access to a page by making the htmlpage -x in the permissions.
Note, do not remove permission from the superdoc display page, or you will cycle into an infinite loop, as the "Permission
Denied" page will refuse to load.  This is comparable to what would theoretically happen if you removed read access from
"Permission Denied" or superdocs in general, so tack that on to "just don't do that" scenarios.


git-svn-id: https://everydevel.svn.sourceforge.net/svnroot/everydevel/trunk@423 b9a9f069-900d-0410-9e71-c61feadb8a10