Completes checkin of groupCache code.  This was inspired by a hack nate put in to Everything2
Inside of nodegroup.pm we now internally cache calls of inGroup() and
inGroupFast()

This fits in to give the system a huge amount of speed, esp when you are
working with a system with a large gods group.

Whenever you want to use inGroup, instead of recursing through the group
it maps it to a hash and then looks up the key.  This has a few minor
side effects:

if you were to:
$grp->insertIntoGroup($FOO);
$grp->inGroup($SOMETHING);
push @{$$grp(group)}, $BAR;
$grp->inGroup($BAR);

the last inGroup call would unexpected turn up as negative.  The reason?
Because you are directly manipulating a copy of $$NODE{group}.  That
should become a fairly opaque structure, and you should only use
insertIntoGroup, removeGroup, and replaceGroup.

Therefore, the above problem (even though it feels a little silly), would
be solved by:

$grp->replaceGroup($$grp{group});

This was chosen as an approach, because it is similar to
selectNodegroupFlat (as that is internally cached inside of
$$NODE{flatgroup});

This will be documented upon the 1.0 release


git-svn-id: https://everydevel.svn.sourceforge.net/svnroot/everydevel/trunk@450 b9a9f069-900d-0410-9e71-c61feadb8a10