This one line fixes solves the problem of these warning messages in the logs:
Warning! Field 'nodelet_group' needs a node of type 'group', but it is pointing to a node of type 'nodeletgroup'!
Warning! Field 'preferred_theme' needs a node of type 'theme', but it is pointing to a node of type 'themesetting'!

(these errors show up onscreen in IIS, a problem that is needs to be resolved)

This traces back to a line swap that needed to happen in Everything::Node::node

Nodes were wound up for revision, but never stored if they didn't need to be.  Therefore we went through the entire
rigamarole of doing the XML::DOM stuff without ever storing it.  XML.pm was throwing these errors.

This keeps the logs cleaner, and improves performance inside of nodes where we do not need to store Revision information.

It is also the more conservative fix, since we could die out before removing the old revisions of the type (for slightly more
perf), but that's an optional enhancement.


git-svn-id: https://everydevel.svn.sourceforge.net/svnroot/everydevel/trunk@458 b9a9f069-900d-0410-9e71-c61feadb8a10