Adds functionality to the nodetype display page that gives you a method map and tells you from which module each method derives from for an easy, simple, and complete list of all of the methods for each nodetype.
In the future if Gigantic Code Lister is ever put back in (or something
similar to it), it will be trivial to link this with a document of that
type of functionality.

* Adds availableMethods as an htmlsnippet. Could be removed and added as
in inline component, but it's quite long, and could make that the
nodetype displaypage confusing.
* Changes nodetype display page to add the availableMethods htmlsnippet
and to clean out the previous similar functionality.


git-svn-id: https://everydevel.svn.sourceforge.net/svnroot/everydevel/trunk@496 b9a9f069-900d-0410-9e71-c61feadb8a10