fix stupid settings bug which has been haunting everyone for the last year
basically, various methods of updating $$USER{vars} rely on the fact that
getHash will always return the same reference which has been stored in $VARS

if it doesn't, then the code will call setVars with its updated vars, and then
this will get clobbered by the code at the end of HTML.pm which writes $VARS
to the database

the problem is, if $$USER{vars} was empty, then an undef was returned instead
of a reference, and HTML.pm clobbered the database with the undef.

the fix is to return the reference and store it in the cache even if $$USER{vars}
is empty, this just involves removing the line which checks for it; i can't see
any reason for this check to exist. tested on a new core, it all works fine.


git-svn-id: https://everydevel.svn.sourceforge.net/svnroot/everydevel/trunk@799 b9a9f069-900d-0410-9e71-c61feadb8a10