Final removal of the last global from Everything::HTML, namely %GLOBAL and associated changes. %GLOBAL was used in part to pass messages from the opcode phase and display-node-selection phase to the nodes to the actual nodes being displayed.
I resolved this problem in two ways.  First, the request object now has a message attribute that
can be set to a message that it deems worthy being given to whoever has the request object.
Second, an argument can be passed to the node object being displayed if the node ends up being a
subroutine. The 'show container' container is passed an argument of the node that it is wrapping
by container.pm

There is also one fix being removal of the $USER from the 'create node' superodc.


git-svn-id: https://everydevel.svn.sourceforge.net/svnroot/everydevel/trunk@997 b9a9f069-900d-0410-9e71-c61feadb8a10