Changes:
NULL values in databases. Changes to ensure that the storage back end can handle null values and that null
values are usually the default in the node and nodetype tables.

When exporting null values in XML an extra tag attribute 'null' is set to 'yes'.  On parsing a tag, the
content is set to undef if this attribute is set

FIX: nodegroup.pm, getNodegroupFlat did not list the nodegroups contained in a nodegroup, only the
contents.  It makes no sense if it doesn't, so this was fixed.