New procedure for installing xml nodes:
Rather than install -1 when there are unknown noderefs, we only install the basic node data i.e.
type_nodetype and 'title'.

Then all nodes in the nodeball are installed.

Once all nodes are installed, we then iterate over the nodes again and fix the noderefs.

This means that we have more flexibility in the backend sql databases that can better store the node in a
way it sees fit rather than having to deal with -1 references that will break things like foreign key
constraints.

TODO: expand this to nodeball updates