First step in sorting out an Apache bug where nodes can't find their node type.
In this commit I'm adding tests to check that 'virtual' nodes can set/get their accessors properly. I've
also removed that code that auto-inserted a location for a node.  That was the wrong place for that logic.
If the frontend wants a node in a certain location, it should put it there and not rely on the backend to
second-guess it.