Change nodetype/metadata hierarchy.
node.pm is a nodetype.pm, but nodetype inheritis its functionality from node.pm.  This can cause all
sorts of problems.  Every solution has had to deal with this circularity.  So does this one.

Because, from a perl package point-of-view, nodetype inherits from node, it makes most logical to
sense, to me anyway, not to fiddle with this hierarchy too much:

- the nodetype nodetype is constructed by pulling the data from the database in the way a normal
nodetype is.
- node.pm's meta data comes from a new class NodetypeMetaData whose attributes come from pulling in the
data that would ordinarily be used to construct the nodetype node.